move UUID/base58 tests into package
